Patched from TightVNC 1.3.10
git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@856 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ vnc_winsrc/vncviewer/ClientConnectionClipboard.cpp
@@ -94,7 +94,7 @@ void ClientConnection::ProcessLocalClipboardChange()
 // we need to copy it into the local clipboard.
 // Called by ClientConnection::ReadServerCutText()
 
-void ClientConnection::UpdateLocalClipboard(char *buf, int len) {
+void ClientConnection::UpdateLocalClipboard(char *buf, size_t len) {
 	
 	if (m_opts.m_DisableClipboard)
 		return;
